+++ NEWS.md
@@ -5,6 +5,7 @@
 - Added `block_size()`, a function to calculate the dimensions and mass of a rip rap block
 - Added `shear_stress()`, a function to calculate shear stress, shear velocity, and dimensionless shear stress
 - Added `wt_sup()`, a function to calculate the superelevation of the water table in a river curve
+- Added `filterlayer()`, a function to calculate the range of the grain size distribution of a filter layer
 
 
 # hydReng 0.1.0